practice assignment for creating routes
