+++ routes.js
@@ -1,0 +1,31 @@
+const fs = require('fs');
+
+const requestHandler = (req, res) => {
+    const url = req.url;
+    const method = req.method;
+    if (url === '/'){
+        res.write('<html>');
+        res.write('<head><title>Greetings</title></body>');
+        res.write('<body><h1>Enter Username</h1><form action="/create-user" method="POST"><input type="text" name="username"><button type="submit">Submit</button></form></body>');
+        res.write('</html>');
+        return res.end();
+    } 
+    if (url === '/create-user' && method === 'POST'){
+        const body = [];
+        req.on('data', (chunk) => {
+            body.push(chunk);
+        });
+        req.on('end', () => {
+            const parsedBody = Buffer.concat(body).toString();
+            const username = parsedBody.split('=')[1];
+            res.write('<html>');
+            res.write('<head><title>Greetings from user</title></body>');
+            res.write(`<body><h1>${username}</h1></body>`);
+            res.write('</html>');
+            res.statusCode = 302;
+            return res.end();
+            });    
+    }
+};
+
+module.exports = requestHandler;
